Grandes alterações de validação de dados

+++ src/main/java/PratoPronto/back/service/AvaliacaoService.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import PratoPronto.back.model.Avaliacao;
+import PratoPronto.back.model.Receita;
 import PratoPronto.back.repository.AvaliacaoRepository;
 
 @Service
@@ -17,13 +18,20 @@ public class AvaliacaoService {
     public List<Avaliacao> listarTodas() {
         return avaliacaoRepository.findAll();
     }
-    
+
+    public List<Avaliacao> listarPorReceita(Receita receita) {
+        return avaliacaoRepository.findByReceita(receita);
+    }
 
     public Avaliacao salvar(Avaliacao avaliacao) {
         return avaliacaoRepository.save(avaliacao);
     }
 
     public void deletar(Long id) {
-        avaliacaoRepository.deleteById(id);
+        if(avaliacaoRepository.existsById(id)) {
+            avaliacaoRepository.deleteById(id);
+        } else {
+            throw new RuntimeException("Avaliação com id " + id + " não encontrada.");
+        }
     }
 }
